+++ packages/deposit-screener/src/waitEstimation/index.ts
@@ -1,7 +1,12 @@
 import { Job, Queue } from "bullmq";
 import { DepositScreenerDB } from "../db";
-import { DepositRequestJobData, DepositRequestStatus } from "../types";
-import { Address, AssetTrait, DepositRequest } from "@nocturne-xyz/sdk";
+import { DepositRequestJobData } from "../types";
+import {
+  Address,
+  AssetTrait,
+  DepositRequest,
+  DepositRequestStatus,
+} from "@nocturne-xyz/sdk";
 import { ScreenerDelayCalculator } from "../screenerDelay";
 import { ScreeningApi } from "../screening";
 import {