+++ packages/e2e-tests/src/utils.ts
@@ -1,4 +1,5 @@
 import {
+  DepositStatusResponse,
   MockSubtreeUpdateProver,
   OperationStatus,
   ProvenOperation,
@@ -15,7 +16,6 @@ import { WasmSubtreeUpdateProver } from "@nocturne-xyz/local-prover";
 import IORedis from "ioredis";
 import { RedisMemoryServer } from "redis-memory-server";
 import { thunk } from "@nocturne-xyz/sdk";
-import { DepositStatusResponse } from "@nocturne-xyz/deposit-screener";
 
 const ROOT_DIR = findWorkspaceRoot()!;
 const EXECUTABLE_CMD = `${ROOT_DIR}/rapidsnark/build/prover`;
@@ -94,7 +94,7 @@ export async function submitAndProcessOperation(
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(op),
+      body: JSON.stringify({ operation: op }),
     });
     const resJson = await res.json();
     console.log("bundler server response: ", resJson);