+++ packages/site/src/config/index.ts
@@ -1,3 +1,3 @@
 export { SNAP_ID } from "./snap";
-export { TELLER_CONTRACT_ADDRESS } from "./contracts";
+export { DEPOSIT_MANAGER_CONTRACT_ADDRESS } from "./contracts";
 export { BUNDLER_ENDPOINT } from "./bundler";